+++ Sources/PeripheryKit/Indexer/SwiftIndexer.swift
@@ -357,6 +357,7 @@ public final class SwiftIndexer {
                     decl.modifiers = Set(result.modifiers)
                     decl.commentCommands = Set(result.commentCommands)
                     decl.declaredType = result.variableType
+                    decl.ifLetShorthandIdentifiers = result.ifLetShorthandIdentifiers
 
                     for ref in decl.references.union(decl.related) {
                         if result.inheritedTypeLocations.contains(ref.location) {